refactor: move useVideoStats to libs/core and fix setState-during-render
- Addresses Copilot thread 21: replace derived-state-during-render pattern
  (setState called inside render body) with useLayoutEffect([publisher]),
  which synchronises stats synchronously before the browser paints and
  avoids potential render loops with non-stable publisher references.
  Removes the trackedPublisher state and the if-block that called setState
  during render.

- Addresses Copilot thread 22: useVideoStats is Vonage SDK-specific but
  Vera-agnostic, so it belongs in libs/core (faceless reusable logic) per
  architecture guidelines. Extracted into:
    libs/core/src/hooks/useVideoStats/
      types.ts                      – VideoStats type
      helpers/readPublisherStats.ts – reads width/height/frameRate from SDK publisher
      helpers/formatResolution.ts   – formats height → "720p"
      helpers/formatFrameRate.ts    – formats frameRate → "30fps"
      helpers/index.ts              – re-exports formatters
      useVideoStats.ts              – the hook
      useVideoStats.spec.ts         – tests (moved from frontend)
      index.ts                      – public barrel export

  VideoStatsOverlay.tsx import updated to @core/hooks/useVideoStats.
  Old frontend/src/hooks/useVideoStats.tsx and its spec deleted.
