+++ libs/core/src/hooks/useVideoStats/helpers/index.ts
@@ -1,0 +1,2 @@
+export { default as formatResolution } from './formatResolution';
+export { default as formatFrameRate } from './formatFrameRate';
