+++ libs/core/src/hooks/useVideoStats/index.ts
@@ -1,0 +1,3 @@
+export { default } from './useVideoStats';
+export { formatResolution, formatFrameRate } from './helpers';
+export type { VideoStats } from './types';
