+++ libs/core/src/hooks/useVideoStats/types.ts
@@ -1,0 +1,5 @@
+export type VideoStats = {
+  width: number | null;
+  height: number | null;
+  frameRate: number | null;
+};
